Added services for User

+++ parking/parking/src/main/java/com/wthing/parking/services/implementations/ReservationServiceImpl.java
@@ -1,6 +1,8 @@
 package com.wthing.parking.services.implementations;
 
 import com.wthing.parking.dto.ReservationDto;
+import com.wthing.parking.enums.ParkingSpotStatusEnum;
+import com.wthing.parking.enums.ReservationStatusEnum;
 import com.wthing.parking.mappers.Mappers;
 import com.wthing.parking.models.ParkingSpot;
 import com.wthing.parking.models.Reservation;
@@ -52,13 +54,55 @@ public class ReservationServiceImpl implements ReservationService {
         ParkingSpot parkingSpot = parkingSpotRepo.findById(reservationDto.getParkingSpotId())
                 .orElseThrow(() -> new IllegalArgumentException(SPOT_NOT_FOUND));
 
-        reservation.setUser(user);
-        reservation.setParkingSpot(parkingSpot);
-        reservation.setStartTime(reservationDto.getStartTime());
-        reservation.setEndTime(reservationDto.getEndTime());
-        reservation.setStatus(reservationDto.getStatus());
+        if (user != null && parkingSpot != null && ParkingSpotStatusEnum.FREE.equals(parkingSpot.getStatus())) {
+            reservation.setUser(user);
+            reservation.setParkingSpot(parkingSpot);
+            reservation.setStartTime(reservationDto.getStartTime());
+            reservation.setEndTime(reservationDto.getEndTime());
+            reservation.setStatus(ReservationStatusEnum.RESERVED);
 
-        return reservationRepo.save(reservation);
+            parkingSpot.setStatus(ParkingSpotStatusEnum.RESERVED);
+            parkingSpotRepo.save(parkingSpot);
+
+            return reservationRepo.save(reservation);
+        }
+        return null;
+    }
+
+    @Override
+    public Reservation startReservation(Long reservationId) {
+        Reservation reservation = reservationRepo.findById(reservationId)
+                .orElseThrow(() -> new IllegalArgumentException(RESERVATION_NOT_FOUND));
+
+        if (reservation != null && ReservationStatusEnum.RESERVED.equals(reservation.getStatus())) {
+            reservation.setStatus(ReservationStatusEnum.ACTIVE);
+
+            ParkingSpot parkingSpot = reservation.getParkingSpot();
+            parkingSpot.setStatus(ParkingSpotStatusEnum.OCCUPIED);
+            parkingSpotRepo.save(parkingSpot);
+
+            return reservationRepo.save(reservation);
+        }
+
+        return null;
+    }
+
+    @Override
+    public Reservation endReservation(Long reservationId) {
+        Reservation reservation = reservationRepo.findById(reservationId)
+                .orElseThrow(() -> new IllegalArgumentException(RESERVATION_NOT_FOUND));
+
+        if (reservation != null && ReservationStatusEnum.ACTIVE.equals(reservation.getStatus())) {
+            reservation.setStatus(ReservationStatusEnum.COMPLETED);
+
+            ParkingSpot parkingSpot = reservation.getParkingSpot();
+            parkingSpot.setStatus(ParkingSpotStatusEnum.FREE);
+            parkingSpotRepo.save(parkingSpot);
+
+            return reservationRepo.save(reservation);
+        }
+
+        return null;
     }
 
     @Override
@@ -71,6 +115,27 @@ public class ReservationServiceImpl implements ReservationService {
 
     @Override
     public void deleteById(Long reservationId) {
-        reservationRepo.deleteById(reservationId);
+        Reservation reservation = reservationRepo.findById(reservationId)
+                .orElseThrow(() -> new IllegalArgumentException(RESERVATION_NOT_FOUND));
+
+        if (reservation != null) {
+            ParkingSpot parkingSpot = reservation.getParkingSpot();
+            if (ReservationStatusEnum.RESERVED.equals(reservation.getStatus())) {
+                reservation.setStatus(ReservationStatusEnum.CANCELED);
+
+                parkingSpot.setStatus(ParkingSpotStatusEnum.FREE);
+                parkingSpotRepo.save(parkingSpot);
+            }
+            reservationRepo.deleteById(reservationId);
+        }
+    }
+
+    @Override
+    public List<ReservationDto> getActiveReservations() {
+        List<Reservation> active = reservationRepo.findByStatus(ReservationStatusEnum.ACTIVE);
+
+        return active.stream()
+                .map(Mappers::mapToReservationDto)
+                .collect(Collectors.toList());
     }
 }
